requirements.txt
Flask
SpeechRecognition
gTTS
PyPDF2
google-generativeai
pyaudio
